% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{absorption}
\alias{absorption}
\title{Absorption}
\usage{
absorption(bString, model = NULL)
}
\arguments{
\item{bString}{a disjunctive normal form or binary vector according to model}

\item{model}{model for respective binary vector}
}
\value{
bString after absorption law
}
\description{
applies absorption law to a disjuncitve normal form
}
\examples{
graph <- c("A+B=C", "A=C")
absorption(graph)
}
\author{
Martin Pirkl
}
