% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnem_main.r
\name{absorptionII}
\alias{absorptionII}
\title{Inverse absorption}
\usage{
absorptionII(bString, model = NULL)
}
\arguments{
\item{bString}{a disjunctive normal form or binary vector according to model}

\item{model}{model for respective binary vector}
}
\value{
bString after "inverse" absorption law
}
\description{
applies "inverse" absorption law to a disjuncitve normal form
}
\examples{
graph <- c("A+B=C", "A=C")
absorptionII(graph)
}
\author{
Martin Pirkl
}
