% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnem_main.r
\name{addNoise}
\alias{addNoise}
\title{Add noise}
\usage{
addNoise(sim, sd = 1)
}
\arguments{
\item{sim}{bnemsim object from simBoolGtn}

\item{sd}{standard deviation for the rnorm function}
}
\value{
noisy fold-change matrix
}
\description{
Adds noise to simulated data
}
\examples{
sim <- simBoolGtn(Sgenes = 10, maxEdges = 10, negation=0.1,layer=1)
fc <- addNoise(sim,sd=1)
}
\author{
Martin Pirkl
}
