% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\docType{data}
\name{bcr}
\alias{bcr}
\title{B-Cell receptor signalling perturbations}
\usage{
bcr
}
\description{
Processed data from experiments with a stimulated B-Cell receptor (bcr)
and perturbed signalling genes. The raw data is available at
https://www.ncbi.nlm.nih.gov/geo/ with accession id GSE68761. For the
process steps we refer to the publication
Martin Pirkl, Elisabeth Hand, Dieter Kube, Rainer Spang, Analyzing
synergistic and non-synergistic interactions in signalling pathways
using Boolean Nested Effect Models, Bioinformatics, Volume 32, Issue 6,
15 March 2016, Pages 893–900, https://doi.org/10.1093/bioinformatics/btv680.
Alternatively see also the function processDataBCR for details.
}
\examples{
data(bcr)
}
\references{
Martin Pirkl, Elisabeth Hand, Dieter Kube, Rainer Spang,
Analyzing synergistic and non-synergistic interactions in signalling
pathways using Boolean Nested Effect Models, Bioinformatics, Volume 32,
Issue 6, 15 March 2016, Pages 893–900,
https://doi.org/10.1093/bioinformatics/btv680
}
