% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{convertGraph}
\alias{convertGraph}
\title{Convert normal form}
\usage{
convertGraph(g)
}
\arguments{
\item{g}{graph in normal form}
}
\value{
converted graph normal form
}
\description{
converts a disjunctive normal form into a conjunctive normal form and
vice versa
}
\examples{
g <- "A+B=C"
g2 <- convertGraph(g)
}
\author{
Martin Pirkl
}
