% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiNEM2Bg.R
\name{epiNEM2Bg}
\alias{epiNEM2Bg}
\title{Convert epiNEM model into general Boolean graph.
Only needed for comparing accuracy of inferred network for bnem and epiNEM.}
\usage{
epiNEM2Bg(t)
}
\arguments{
\item{t}{full epiNEM model}
}
\value{
boolean hyper-graph
}
\description{
Convert epiNEM model into general Boolean graph.
Only needed for comparing accuracy of inferred network for bnem and epiNEM.
}
\examples{
library(epiNEM)
topology <- CreateTopology(3, 1, force = TRUE)
topology <- unlist(unique(topology), recursive = FALSE)
extTopology <- ExtendTopology(topology$model, 100)
b <- EpiNEM2BooleanGraph(extTopology)
}
\author{
Martin Pirkl
}
\seealso{
CreateTopology
}

