% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{plot.bnemsim}
\alias{plot.bnemsim}
\title{plot simulation object}
\usage{
plot.bnemsim(x, ...)
}
\arguments{
\item{x}{bnemsim object}

\item{...}{further arguments; see function mnem::plotDnf}
}
\value{
plot of boolean network
}
\description{
plots the boolen network as disjunctive normal form
}
\examples{
sim <- simBoolGtn()
plot(sim)
}
\author{
Martin Pirkl
}
