% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{processData}
\alias{processData}
\title{BCR perturbation reproduction}
\usage{
processData(path = "")
}
\arguments{
\item{path}{path to the CEL.gz files}
}
\value{
data matrix with foldchanges
}
\description{
Produce the application data from the BCR paper
of Pirkl, et al., 2016, Bioinformatics. Raw data is available at
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE68761
}
\examples{
\dontrun{
processData()
}
data(bcr)
}
\author{
Martin Pirkl
}
