% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnem_main.r
\name{processDataBCR}
\alias{processDataBCR}
\title{BCR perturbation reproduction}
\usage{
processDataBCR(path = "", combsign = FALSE)
}
\arguments{
\item{path}{path to the CEL.gz/Cel files}

\item{combsign}{if TRUE includes all covariates in ComBat analysis
to estimate batch effects.}
}
\value{
list with the full foldchanges and epxression matrix,
a reduced foldchange matrix and the design matrix for the computations
}
\description{
Produce the application data from the BCR paper
of Pirkl, et al., 2016, Bioinformatics. Raw data is available at
https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE68761
}
\examples{
\dontrun{
processDataBCR()
}
data(bcr)
}
\author{
Martin Pirkl
}
