% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnem_main.r
\name{randomDnf}
\alias{randomDnf}
\title{sample normal form}
\usage{
randomDnf(
  vertices = 10,
  negation = TRUE,
  max.edge.size = NULL,
  max.edges = NULL,
  dag = FALSE
)
}
\arguments{
\item{vertices}{number of vertices}

\item{negation}{if TRUE, negations (NOT gates) are allowed}

\item{max.edge.size}{maximal number of inputs per edge}

\item{max.edges}{maximal number of hyper-edges}

\item{dag}{if TRUE, graph will be acyclic}
}
\value{
random hyper-graph in normal form
}
\description{
creates a random normal form or hyper-graph
}
\examples{
g <- randomDnf(10)
}
\author{
Martin Pirkl
}
