% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduceGraph.R
\name{reduceGraph}
\alias{reduceGraph}
\title{reduces the size of a graph if ossible to an equivalent sub-graph}
\usage{
reduceGraph(bString, model, CNOlist)
}
\arguments{
\item{bString}{binary vector indicating the sub-graph given a model}

\item{model}{model object for the whole graph space}

\item{CNOlist}{CNOlist object}
}
\value{
equivalent sub-graph denoted by bString
}
\description{
reduces the size of a graph if ossible to an equivalent sub-graph
}
\examples{
library(bnem)
sifMatrix <- rbind(c("A", 1, "B"), c("A", 1, "C"), c("B", 1, "D"), c("C", 1, "D"))
write.table(sifMatrix, file = "temp.sif", sep = "\\t", row.names = FALSE, col.names = FALSE,
quote = FALSE)
PKN <- readSIF("temp.sif")
unlink('temp.sif')
model <- preprocessing(CNOlist, PKN, maxInputsPerGate = 100)
CNOlist <- dummyCNOlist("A", c("B","C","D"), maxStim = 1, maxInhibit = 2, signal = c("A", "B","C","D"))
bString <- reduceGraph(rep(1, length(model$reacID)), model, CNOlist)
}
\author{
Martin Pirkl
}

