% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{scoreDnf}
\alias{scoreDnf}
\title{score a boolean network}
\usage{
scoreDnf(bString, CNOlist, fc, model, method = "llr")
}
\arguments{
\item{bString}{binary string denoting the boolean network}

\item{CNOlist}{CNOlist object}

\item{fc}{data (see ?bnem)}

\item{model}{network model}

\item{method}{see ?bnem}
}
\value{
numeric value (score)
}
\description{
computes the score of a boolean network given the model and data
}
\examples{
sim <- simBoolGtn()
scoreDnf(sim$bString, sim$CNOlist, sim$fc, sim$model)
}
\author{
Martin Pirkl
}
