% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{transClose}
\alias{transClose}
\title{transitive closure}
\usage{
transClose(g, max.iter = NULL, verbose = FALSE)
}
\arguments{
\item{g}{hyper-graph in normal form}

\item{max.iter}{maximal iteration till convergence}

\item{verbose}{verbose output?}
}
\value{
transitive closure in normal form
}
\description{
calculates transitive closure of a hyper-graph
}
\examples{
g <- c("A=B", "B=C")
gclose <- transClose(g)
}
\author{
Martin Pirkl
}
