% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transRed.R
\name{transRed}
\alias{transRed}
\title{calculates transitive reduciton of a hyper-graph in normal form}
\usage{
transRed(g, max.iter = NULL, verbose = FALSE)
}
\arguments{
\item{g}{hyper-graph in normal form}

\item{max.iter}{maximal number of iterations till convergence}

\item{verbose}{verbose output?}
}
\value{
transitive reduction of the hyper-graph in normal form
}
\description{
calculates transitive reduciton of a hyper-graph in normal form
}
\examples{
library(bnem)
g <- c("A=B", "A=C", "B=C", "B=D", "!A=D")
gred <- transRed(g)
}
\author{
Martin Pirkl
}

