% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnem_main.r
\name{validateGraph}
\alias{validateGraph}
\title{validate graph}
\usage{
validateGraph(
  CNOlist,
  fc = NULL,
  exprs = NULL,
  approach = "fc",
  model,
  bString,
  Egenes = 25,
  Sgene = 1,
  parameters = list(cutOffs = c(0, 1, 0), scoring = c(0.1, 0.2, 0.9)),
  plot = TRUE,
  disc = 0,
  affyIds = TRUE,
  relFit = FALSE,
  xrot = 25,
  Rowv = FALSE,
  Colv = FALSE,
  dendrogram = "none",
  soft = TRUE,
  colSideColors = NULL,
  affychip = "hgu133plus2",
  method = "s",
  ranks = FALSE,
  breaks = NULL,
  col = "RdYlGn",
  sizeFac = 10^-10,
  verbose = TRUE,
  order = "rank",
  ...
)
}
\arguments{
\item{CNOlist}{CNOlist object.}

\item{fc}{matrix of foldchanges (observed response scheme, ORS).}

\item{exprs}{optional matrix of normalized expression (observed
activation scheme).}

\item{approach}{default "fc" for foldchanges or signed effect probabilities,
"abs" for absolute effects or probabilities}

\item{model}{Model object.}

\item{bString}{Binary string denoting the hyper-graph.}

\item{Egenes}{Maximal number of visualized E-genes.}

\item{Sgene}{Integer denoting the S-gene. See
colnames(CNOlist@signals[[1]]) to match integer with S-gene name.}

\item{parameters}{see ?bnem}

\item{plot}{Plot the heatmap. If FALSE, only corresponding
information is printed.}

\item{disc}{Discretize the data.}

\item{affyIds}{Experimental. Turn Affymetrix Ids into HGNC
gene symbols.}

\item{relFit}{see ?bnem}

\item{xrot}{See package ?epiNEM::HeatmapOP}

\item{Rowv}{See package ?epiNEM::HeatmapOP}

\item{Colv}{See package ?epiNEM::HeatmapOP}

\item{dendrogram}{See package ?epiNEM::HeatmapOP}

\item{soft}{weighting the expected pattern, if TRUE}

\item{colSideColors}{See package ?epiNEM::HeatmapOP}

\item{affychip}{Define Affymetrix chip used to generate the data
(optional).}

\item{method}{see ?bnem}

\item{ranks}{Turns data into ranks, if TRUE}

\item{breaks}{See package ?epiNEM::HeatmapOP}

\item{col}{See package ?epiNEM::HeatmapOP}

\item{sizeFac}{size factor penelizing the hyper-graph size.}

\item{verbose}{verbose output, if TRUE}

\item{order}{Order by "rank" or "name" or "none"}

\item{...}{additional arguments for ?epiNEM::HeatmapOP}
}
\value{
lattice object with matrix information
}
\description{
plotting the observed response scheme of an effect reporter and the
expected response scheme of the regulating signalling gene
}
\examples{
sifMatrix <- rbind(c("A", 1, "B"), c("A", 1, "C"), c("B", 1, "D"),
c("C", 1, "D"))
write.table(sifMatrix, file = "temp.sif", sep = "\t", row.names = FALSE,
col.names = FALSE,
quote = FALSE)
PKN <- CellNOptR::readSIF("temp.sif")
unlink('temp.sif')
CNOlist <- dummyCNOlist("A", c("B","C","D"), maxStim = 1, maxInhibit = 2,
signal = c("A", "B","C","D"))
model <- CellNOptR::preprocessing(CNOlist, PKN, maxInputsPerGate = 100)
exprs <- matrix(rnorm(nrow(slot(CNOlist, "cues"))*10), 10,
nrow(slot(CNOlist, "cues")))
fc <- computeFc(CNOlist, exprs)
initBstring <- rep(0, length(model$reacID))
res <- bnem(search = "greedy", CNOlist = CNOlist, fc = fc,
model = model, parallel = NULL, initBstring = initBstring, draw = FALSE,
verbose = FALSE, maxSteps = Inf)
rownames(fc) <- seq_len(nrow(fc))
val <- validateGraph(CNOlist = CNOlist, fc = fc, model = model,
bString = res$bString, Egenes = 10, Sgene = 4)
}
\author{
Martin Pirkl
}
