% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGraph.R
\name{convertGraph}
\alias{convertGraph}
\title{converts a disjunctive normal form into a conjunctive normal form and vice versa}
\usage{
convertGraph(g)
}
\arguments{
\item{g}{graph in normal form}
}
\value{
converted graph normal form
}
\description{
converts a disjunctive normal form into a conjunctive normal form and vice versa
}
\examples{
library(bnem)
g <- "A+B=C"
g2 <- convertGraph(g)
}
\author{
Martin Pirkl
}

