% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnem_main.r
\name{convertGraph}
\alias{convertGraph}
\title{Convert normal form}
\usage{
convertGraph(g)
}
\arguments{
\item{g}{graph in normal form}
}
\value{
converted graph normal form
}
\description{
converts a disjunctive normal form into a conjunctive normal form and
vice versa;
input graph as disjunctive normal form like that:
c("A+B=D", "C=D", "G+F=U", ...); output is the dual element
also in disjunctive normal form;
}
\examples{
g <- "A+B=C"
g2 <- convertGraph(g)
}
\author{
Martin Pirkl
}
