% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummyCNOlist.R
\name{dummyCNOlist}
\alias{dummyCNOlist}
\title{creates a general CNOlist object from meta information}
\usage{
dummyCNOlist(stimuli = NULL, inhibitors = NULL, maxStim = 0,
  maxInhibit = 0, signals = NULL)
}
\arguments{
\item{stimuli}{character vector of stimulated genes}

\item{maxStim}{maximal number of stimulated genes for a single experiment}

\item{maxInhibit}{maximal number of inhibited genes for a single experiment}

\item{signals}{character vector of genes which can directly regulate effect reporters}

\item{character}{vector of inhibited genes}
}
\value{
general CNOlist object
}
\description{
creates a general CNOlist object from meta information
}
\examples{
sifMatrix <- rbind(c("A", 1, "B"), c("A", 1, "C"), c("B", 1, "D"), c("C", 1, "D"))
write.table(sifMatrix, file = "temp.sif", sep = "\\t", row.names = FALSE, col.names = FALSE,
quote = FALSE)
PKN <- readSIF("temp.sif")
unlink('temp.sif')
CNOlist <- dummyCNOlist("A", c("B","C","D"), maxStim = 1, maxInhibit = 2, signals = c("A", "B","C","D"))
}
\author{
Martin Pirkl
}

