% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnem_main.r
\name{plot.bnem}
\alias{plot.bnem}
\title{plot bnem opbject}
\usage{
\method{plot}{bnem}(x, ...)
}
\arguments{
\item{x}{bnemsim object}

\item{...}{further arguments; see function mnem::plotDnf}
}
\value{
plot of boolean network
}
\description{
plots the boolen network as disjunctive normal form
}
\examples{
sim <- simBoolGtn()
plot(sim)
}
\author{
Martin Pirkl
}
