% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomDnf.R
\name{randomDnf}
\alias{randomDnf}
\title{creates a random normal form or hyper-graph}
\usage{
randomDnf(vertices = 10, negation = TRUE, max.edge.size = NULL,
  max.edges = NULL, dag = FALSE)
}
\arguments{
\item{negation}{allowed?}

\item{max.edge.size}{maximal number of inputs per edge}

\item{max.edges}{maximal number of hyper-edges}

\item{dag}{is the graph to ba a dag?}

\item{maximal}{number of vertices}
}
\value{
random hyper-graph in normal form
}
\description{
creates a random normal form or hyper-graph
}
\examples{
library(bnem)
g <- randomDnf(10)
}
\author{
Martin Pirkl
}

