% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{reduceGraph}
\alias{reduceGraph}
\title{reduce graph}
\usage{
reduceGraph(bString, model, CNOlist)
}
\arguments{
\item{bString}{binary vector indicating the sub-graph given a model}

\item{model}{model object for the whole graph space}

\item{CNOlist}{CNOlist object}
}
\value{
equivalent sub-graph denoted by bString
}
\description{
reduces the size of a graph if possible to an equivalent sub-graph
}
\examples{
library(CellNOptR)
sifMatrix <- rbind(c("A", 1, "B"), c("A", 1, "C"), c("B", 1, "D"),
c("C", 1, "D"))
write.table(sifMatrix, file = "temp.sif", sep = "\\t", row.names = FALSE,
col.names = FALSE,
quote = FALSE)
PKN <- CellNOptR::readSIF("temp.sif")
unlink('temp.sif')
CNOlist <- dummyCNOlist("A", c("B","C","D"), maxStim = 1, maxInhibit = 2,
signal = c("A", "B","C","D"))
model <- CellNOptR::preprocessing(CNOlist, PKN, maxInputsPerGate = 100)
bString <- reduceGraph(rep(1, length(model$reacID)), model, CNOlist)
}
\author{
Martin Pirkl
}
