% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnem_main.r
\name{scoreDnf}
\alias{scoreDnf}
\title{score a boolean network}
\usage{
scoreDnf(
  bString,
  CNOlist,
  fc,
  model,
  method = "llr",
  sizeFac = 10^-10,
  NAFac = 1,
  parameters = list(cutOffs = c(0, 1, 0), scoring = c(0.25, 0.5, 2)),
  approach = "fc",
  NEMlist = NULL,
  relFit = FALSE,
  opt = "min",
  verbose = FALSE
)
}
\arguments{
\item{bString}{binary string denoting the boolean network}

\item{CNOlist}{CNOlist object}

\item{fc}{data (see ?bnem)}

\item{model}{network model}

\item{method}{see ?bnem}

\item{sizeFac}{numeric penalty for network size}

\item{NAFac}{numeric penalty for network size}

\item{parameters}{parameters for discrete case (not recommended);
has to ba list with entries cutOffs and scoring:
cutOffs = c(a,b,c) with a (cutoff for real zeros),
b (cutoff for real effects),
c = -1 for normal scoring, c between 0 and
1 for keeping only relevant % of E-genes,
between -1 and 0 for keeping only a specific quantile of E-genes,
and c > 1 for keeping the top c E-genes;
scoring = c(a,b,c) with a (weight for real effects),
c (weight for real zeros),
b (multiplicator for effects/zeros between a and c);}

\item{approach}{"fc" for log foldchanges, "exprs" for
absolute expression values (not recommended)}

\item{NEMlist}{NEMlist object (optional)}

\item{relFit}{if TRUE returns a network specific relative score}

\item{opt}{"min" returns minimal optimum, "max" return 1-minimum}

\item{verbose}{if TRUE returns verbose output}
}
\value{
numeric value (score)
}
\description{
computes the score of a boolean network given the model and data
}
\examples{
sim <- simBoolGtn()
scoreDnf(sim$bString, sim$CNOlist, sim$fc, sim$model)
}
\author{
Martin Pirkl
}
