% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{simBoolGtn}
\alias{simBoolGtn}
\title{Sample random network and simulate data}
\usage{
simBoolGtn(n = 10, e = 25, s = 2, dag = TRUE, maxStim = 2,
  maxInhibit = 1, m = 10, mflip = 0.33, r = 3, sd = 1,
  keepsif = FALSE, maxcount = 10, negation = TRUE, allstim = FALSE)
}
\arguments{
\item{n}{number of S-genes}

\item{e}{number of maximum edges}

\item{s}{number of stimulated S-genes}

\item{dag}{if TRUE graph will be acyclic}

\item{maxStim}{maximum stimulated S-genes in the data samples}

\item{maxInhibit}{maximum inhibited number of S-genes in the data samples}

\item{m}{E-genes per S-gene}

\item{mflip}{number of inhibited E-genes}

\item{r}{numbero f replicates}

\item{sd}{standard deviation for the gaussian noise}

\item{keepsif}{if TRUE does not delete sif file, which encodes the network}

\item{maxcount}{while loopes ensure a reasonable network, maxcount makes sure while loops do not run in infinity}

\item{negation}{if TRUE negative edges are allowed}

\item{allstim}{full network in which all S-genes are possibly stimulated}
}
\value{
list with the corresponding prior graph, ground truth network and data
}
\description{
Draw a random prior network, samples a ground truth from the full boolean extension and generates data
}
\examples{
sim <- simBoolGtn()
}
\author{
Martin Pirkl
}
