% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{simBoolGtn}
\alias{simBoolGtn}
\title{Sample random network and simulate data}
\usage{
simBoolGtn(Sgenes = 10, maxEdges = 25, stimGenes = 2, prob = 1,
  dag = TRUE, maxSize = 2, maxStim = 2, maxInhibit = 1,
  Egenes = 10, flip = 0.33, reps = 3, sd = 1, keepsif = FALSE,
  maxcount = 10, negation = TRUE, allstim = FALSE, verbose = FALSE)
}
\arguments{
\item{Sgenes}{number of S-genes}

\item{maxEdges}{number of maximum edges in the DAG}

\item{stimGenes}{number of stimulated S-genes}

\item{prob}{scaling factor for uniform samplign of a probability to sample
children in the next hierarchical layer}

\item{dag}{if TRUE graph will be acyclic}

\item{maxSize}{maximum number of S-genes in a disjunction or clause}

\item{maxStim}{maximum stimulated S-genes in the data samples}

\item{maxInhibit}{maximum inhibited S-genes in the data samples}

\item{Egenes}{E-genes per S-gene}

\item{flip}{number of inhibited E-genes}

\item{reps}{numbero f replicates}

\item{sd}{standard deviation for the gaussian noise}

\item{keepsif}{if TRUE does not delete sif file, which encodes the network}

\item{maxcount}{while loopes ensure a reasonable network, maxcount makes sure
while loops do not run in infinity}

\item{negation}{if TRUE negative edges are allowed}

\item{allstim}{full network in which all S-genes are possibly stimulated}

\item{verbose}{TRUE for verbose output}
}
\value{
list with the corresponding prior graph, ground truth network and
data example
}
\description{
Draw a random prior network, samples a ground truth from the full boolean
extension and generates data
}
\examples{
sim <- simBoolGtn()
plot(sim)
}
\author{
Martin Pirkl
}
