% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bnem_main.r
\name{simBoolGtn}
\alias{simBoolGtn}
\title{Sample random network and simulate data}
\usage{
simBoolGtn(
  Sgenes = 10,
  maxEdges = 25,
  stimGenes = 2,
  layer = 1,
  frac = 0.1,
  maxInDeg = 2,
  dag = TRUE,
  maxSize = 2,
  maxStim = 2,
  maxInhibit = 1,
  Egenes = 10,
  flip = 0.33,
  reps = 1,
  keepsif = FALSE,
  maxcount = 10,
  negation = 0.25,
  allstim = FALSE,
  and = 0.25,
  positive = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{Sgenes}{number of S-genes}

\item{maxEdges}{number of maximum edges in the DAG}

\item{stimGenes}{number of stimulated S-genes}

\item{layer}{scaling factor for the sampling of next Sgene layer; high
means a higher probability for less Sgenes and low more Sgenes}

\item{frac}{fraction of hyper-edges in the ground truth}

\item{maxInDeg}{maximum number of incoming hyper-edges}

\item{dag}{if TRUE graph will be acyclic}

\item{maxSize}{maximum number of S-genes in a disjunction or clause}

\item{maxStim}{maximum of stimulated S-genes in the data samples}

\item{maxInhibit}{maximum of inhibited S-genes in the data samples}

\item{Egenes}{number of E-genes per S-gene, e.g. 10 S-genes and 10
E-genes will return 100 E-genes overall}

\item{flip}{number of inhibited E-genes}

\item{reps}{number of replicates}

\item{keepsif}{if TRUE does not delete sif file, which encodes
the prior network}

\item{maxcount}{while loopes ensure a reasonable network, maxcount makes sure
while loops do not run to infinity}

\item{negation}{has to be greater or equal to 0 and less than 1; if
greater than 0, negation is the sample probability for negative edges}

\item{allstim}{full network in which all S-genes are possibly stimulated}

\item{and}{probability for and gates in the GTN}

\item{positive}{if TRUE, sets all stimulation edges to activation, else
samples inhibitory edges by negation probability}

\item{verbose}{TRUE for verbose output}
}
\value{
list with the corresponding prior graph, ground truth network and
data
}
\description{
Draws a random prior network, samples a ground truth from the full boolean
extension and generates data
}
\examples{
sim <- simBoolGtn()
plot(sim)
}
\author{
Martin Pirkl
}
