% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.r
\name{validateGraph}
\alias{validateGraph}
\title{validate graph}
\usage{
validateGraph(CNOlist, fc = NULL, exprs = NULL, approach = "fc",
  model, bString, Egenes = 25, Sgene = 1, parameters = list(cutOffs =
  c(0, 1, 0), scoring = c(0.1, 0.2, 0.9)), plot = TRUE, disc = 0,
  affyIds = TRUE, sim = 0, relFit = FALSE, complete = FALSE,
  xrot = 25, Rowv = FALSE, Colv = FALSE, dendrogram = "none",
  soft = TRUE, colSideColors = NULL, affychip = "hgu133plus2",
  method = "s", ranks = FALSE, breaks = NULL, col = "RdYlGn",
  csc = TRUE, sizeFac = 10^-10, verbose = TRUE, order = "rank",
  colnames = "bio", ...)
}
\arguments{
\item{CNOlist}{CNOlist object.}

\item{fc}{matrix of foldchanges (observed response scheme, ORS).}

\item{exprs}{optional matrix of normalized expression (observed
activation scheme).}

\item{approach}{not used}

\item{model}{Model object.}

\item{bString}{Binary string denoting the hyper-graph.}

\item{Egenes}{Maximal number of visualized E-genes.}

\item{Sgene}{Integer denoting the S-gene. See
colnames(CNOlist@signals[[1]]) to match integer with S-gene name.}

\item{parameters}{not used}

\item{plot}{Plot the heatmap. If FALSE, only corresponding
information is outputed.}

\item{disc}{Discretize the data.}

\item{affyIds}{Experimental. Turn Affymetrix Ids into HGNC
gene symbols.}

\item{sim}{not used}

\item{relFit}{not used}

\item{complete}{not used}

\item{xrot}{See package epiNEM}

\item{Rowv}{See package epiNEM}

\item{Colv}{See package epiNEM}

\item{dendrogram}{See package epiNEM}

\item{soft}{not used}

\item{colSideColors}{See package epiNEM}

\item{affychip}{Define Affymetrix chip used to generate the data
(optional).}

\item{method}{Scoring method can be a correlation or distance measure.
See ?cor and ?dist for details.}

\item{ranks}{Turn data into ranks.}

\item{breaks}{See package epiNEM}

\item{col}{See package epiNEM}

\item{csc}{not used (devel)}

\item{sizeFac}{Size factor penelizing the hyper-graph size.}

\item{verbose}{Verbose output.}

\item{order}{Order by "rank" or "name".}

\item{colnames}{not used (devel)}

\item{...}{additional arguments}
}
\value{
lattice object with matrix information
}
\description{
plotting the observed response scheme of an effect reporter and the
expected response scheme of the regulating signalling gene
}
\examples{
library(CellNOptR)
sifMatrix <- rbind(c("A", 1, "B"), c("A", 1, "C"), c("B", 1, "D"),
c("C", 1, "D"))
write.table(sifMatrix, file = "temp.sif", sep = "\\t", row.names = FALSE,
col.names = FALSE,
quote = FALSE)
PKN <- readSIF("temp.sif")
unlink('temp.sif')
CNOlist <- dummyCNOlist("A", c("B","C","D"), maxStim = 1, maxInhibit = 2,
signal = c("A", "B","C","D"))
model <- preprocessing(CNOlist, PKN, maxInputsPerGate = 100)
exprs <- matrix(rnorm(nrow(slot(CNOlist, "cues"))*10), 10,
nrow(slot(CNOlist, "cues")))
fc <- computeFc(CNOlist, exprs)
initBstring <- rep(0, length(model$reacID))
res <- bnem(search = "greedy", CNOlist = CNOlist, fc = fc,
model = model, parallel = NULL, initBstring = initBstring, draw = FALSE,
verbose = FALSE, maxSteps = Inf)
rownames(fc) <- seq_len(nrow(fc))
val <- validateGraph(CNOlist = CNOlist, fc = fc, model = model,
bString = res$bString, Egenes = 10, Sgene = 4)
}
\author{
Martin Pirkl
}
